requirements.txt
rich
python-dotenv
openai
python-telegram-bot
requests